Issue/admin dashboard

+++ lib/repositories/utente_repository.dart
@@ -5,6 +5,8 @@ import 'package:resqpet/dao/utente_dao.dart';
 import 'package:resqpet/models/utente.dart';
 import 'package:resqpet/services/auth_service.dart';
 
+import 'package:cloud_functions/cloud_functions.dart';
+
 class UtenteRepository {
   final AuthService _authService;
   final UtenteDao _utenteDao;
@@ -161,10 +163,25 @@ class UtenteRepository {
       throw StateError('Nessun utente autenticato');
     }
 
-    await _utenteDao.deleteById(currentUser.uid);
-    await currentUser.delete();
+    await cancellaAccountById(currentUser.uid);
   }
 
+
+  Future<void> cancellaAccountById(String id) async {
+    final currentUser = _authService.currentUser;
+    if (currentUser == null) {
+      throw StateError('Nessun utente autenticato');
+    }
+
+    await _utenteDao.deleteById(id);
+    await currentUser.delete();
+
+    final callable = FirebaseFunctions.instance
+      .httpsCallable("deleteUserAccountByUID");
+
+    await callable.call({ 'uid': id });
+  }
+  
   Future<Utente> _registraUtente({
     required String email,
     required String password,
@@ -218,5 +235,14 @@ class UtenteRepository {
     );
     
     return await _utenteDao.create(utente);
+  }
+
+  Stream<List<Utente>> getAllExceptAdmin() {
+    return _utenteDao.findAllStream()
+      .map(
+        (utenti) => 
+          utenti.where((utente) => utente.tipo != TipoUtente.admin)
+            .toList()
+      );
   }
 }
